+++ apps/api/src/shared/queue/base-queue.ts
@@ -2,6 +2,9 @@ import { JobsOptions, Queue, QueueEvents, Worker } from 'bullmq';
 import Redis from 'ioredis';
 import { queueConnection } from './queue-connection';
 import { logger } from '../logger/logger';
+import { BullAdapter } from '@bull-board/api/bullAdapter';
+
+export const allQueues = [];
 
 export class BaseQueue {
   queue: Queue;
@@ -14,6 +17,9 @@ export class BaseQueue {
     });
     this.queueEvents = new QueueEvents(name);
     this.worker = worker;
+
+    // push all queue for the job board
+    allQueues.push(new BullAdapter(this.queue, { allowRetries: false }));
 
     worker.on('completed', (job) => {
       logger.info(`Job Completed`, {
@@ -32,7 +38,7 @@ export class BaseQueue {
     });
   }
 
-  async add(name, data, options: JobsOptions = {}) {
+  protected async add(name, data, options: JobsOptions = {}) {
     const job = await this.queue.add(name, data, options);
     return { job, queueEvents: this.queueEvents };
   }