+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [3.4.0](https://github.com/vishal2457/mx/compare/v3.3.0...v3.4.0) (2024-06-14)
+
+
+### Features
+
+* confirmation modal added ([4d5d22a](https://github.com/vishal2457/mx/commit/4d5d22aea345f9eba7c70e9dda8dbbdcaa5c5fc1)), closes [#30](https://github.com/vishal2457/mx/issues/30)
+
+# [3.3.0](https://github.com/vishal2457/mx/compare/v3.2.0...v3.3.0) (2024-06-14)
+
+
+### Features
+
+* inactive match logic added, ([ad5649e](https://github.com/vishal2457/mx/commit/ad5649e549965143a473785b79a5b60f4ee1ebb5))
+
 # [3.2.0](https://github.com/vishal2457/mx/compare/v3.1.0...v3.2.0) (2024-06-13)